+++ app/Domain/Flowers/Models/Flower.php
@@ -5,6 +5,7 @@ namespace App\Domain\Flowers\Models;
 use App\Domain\Rooms\Models\Room;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Database\Factories\FlowerFactory;
 
 class Flower extends Model
 {
@@ -21,5 +22,10 @@ class Flower extends Model
     public function room()
     {
         return $this->belongsTo(Room::class);
+    }
+
+    protected static function newFactory()
+    {
+        return new FlowerFactory();
     }
 }
