--- app/Domain/Rooms/Models/Factories/RoomFactory.php
@@ -1,24 +1,0 @@
-<?php
-
-namespace App\Domain\Rooms\Models\Factories;
-
-use Illuminate\Database\Eloquent\Factories\Factory;
-
-/**
- * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\Model>
- */
-class RoomFactory extends Factory
-{
-    /**
-     * Define the model's default state.
-     *
-     * @return array<string, mixed>
-     */
-    public function definition()
-    {
-        return [
-            'name' => $this->faker->word,
-            'capacity' => $this->faker->numberBetween(1, 10),
-        ];
-    }
-}
