+++ app/Domain/Rooms/Models/Room.php
@@ -5,6 +5,7 @@ namespace App\Domain\Rooms\Models;
 use App\Domain\Flowers\Models\Flower;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Database\Factories\RoomFactory;
 
 class Room extends Model
 {
@@ -18,5 +19,10 @@ class Room extends Model
     public function flowers()
     {
         return $this->hasMany(Flower::class);
+    }
+
+    protected static function newFactory()
+    {
+        return new RoomFactory();
     }
 }
