+++ database/factories/RoomFactory.php
@@ -1,14 +1,16 @@
 <?php
 
-namespace App\Domain\Flowers\Models\Factories;
+namespace Database\Factories;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
+use App\Domain\Rooms\Models\Room;
 
 /**
  * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\Model>
  */
-class FlowerFactory extends Factory
+class RoomFactory extends Factory
 {
+    protected $model = Room::class;
     /**
      * Define the model's default state.
      *
@@ -18,10 +20,7 @@ class FlowerFactory extends Factory
     {
         return [
             'name' => $this->faker->word,
-            'price' => $this->faker->randomFloat(2, 0, 100),
-            'watering_time' => $this->faker->dateTimeBetween('-1 week', 'now'),
-            'watering_interval' => $this->faker->numberBetween(1, 10),
-            'room_id' => $this->faker->numberBetween(1, 10),
+            'capacity' => $this->faker->numberBetween(1, 10),
         ];
     }
 }
